requirements.txt
appnope==0.1.4
asttokens==2.4.1
batch-face==1.5.0
blinker==1.8.2
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
contourpy==1.3.0
coverage==7.6.3
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.9
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.9.0
ftfy==6.3.1
future==1.0.0
icatcher==0.2.3
idna==3.7
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multiprocess==0.70.17
nest_asyncio==1.6.0
networkx==3.2.1
numpy==2.0.1
opencv-python==4.10.0.84
opencv-transforms==0.0.6
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathos==0.3.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sixdrepnet==0.1.6
stack-data==0.6.2
sympy==1.13.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
zipp==3.20.2
